Moved solver module to interfaces
