+++ include/solver/interface.hpp
@@ -1,0 +1,7 @@
+#ifndef _SOLVER_INTERFACE_HPP
+#define _SOLVER_INTERFACE_HPP
+
+#include "solver.hpp"
+#include "time_marching.hpp"
+
+#endif
